randomCommit
