+++ src/index.js
@@ -2,31 +2,23 @@ import React from "react";
 import ReactDOM from "react-dom/client";
 import "./index.css";
 import App from "./App";
-import { BrowserRouter, Routes, Route } from "react-router-dom";
-import ExpensesPage from "./pages/ExpensesPage";
-import { InvoicesPage } from "./pages/InvoicesPage";
-import TestPage from "./pages/TestPage";
+import { BrowserRouter } from "react-router-dom";
+import { ThemeProvider, createTheme } from "@mui/material";
+import CssBaseline from "@mui/material/CssBaseline";
 
+const darkTheme = createTheme({
+  palette: {
+    mode: "dark",
+  },
+});
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
   <React.StrictMode>
     <BrowserRouter>
-      <Routes>
-        <Route path="/" element={<App />}>
-          <Route path="expenses" element={<ExpensesPage />} />
-          <Route path="invoices" element={<InvoicesPage />} />
-          <Route path="test" element={<TestPage />} />
-
-          <Route
-            path="*"
-            element={
-              <main style={{ padding: "1rem" }}>
-                <p> There is nothing here!</p>
-              </main>
-            }
-          />
-        </Route>
-      </Routes>
+      <ThemeProvider theme={darkTheme}>
+        <CssBaseline />
+        <App />
+      </ThemeProvider>
     </BrowserRouter>
   </React.StrictMode>
 );
