+++ src/layouts/Layout.js
@@ -1,6 +1,6 @@
 import React from "react";
-import { Outlet, Link } from "react-router-dom";
-import Navigation from "../components/Navigation";
+import { Outlet } from "react-router-dom";
+import SearchAppBar from "../components/Appbar";
 import Stack from "@mui/material/Stack";
 import Box from "@mui/material/Box";
 
@@ -9,10 +9,10 @@ export const Layout = () => {
     <Stack
       sx={{
         minHeight: "100vh",
-        backgroundColor: (theme) => theme.palette.primary.main,
+        // backgroundColor: (theme) => theme.palette.primary.main,
       }}
     >
-      <Navigation />
+      <SearchAppBar title="Job Searching" />
       <Outlet />
       <Box sx={{ flexGrow: 1 }} />
     </Stack>
